Move watchdog thread into Watchdog.start_thread(), one shared thread for all repos
Addresses review feedback on #358. The standalone _start_watchdog_thread
function in server.py and its _WATCHDOG_INTERVAL constant move into
Watchdog.start_thread() so that Watchdog is self-contained. server.run()
now injects _Watchdog=Watchdog and calls .start_thread() — one daemon
thread watching all repos together. Thread tests move from test_server.py
to test_watchdog.py where they belong.
